+++ lib/src/platform/web/file_picker_web.dart
@@ -6,6 +6,8 @@ import 'package:file_picker/src/api/file_picker_types.dart';
 import 'package:file_picker/src/api/file_picker_result.dart';
 import 'package:file_picker/src/api/platform_file.dart';
 import 'package:file_picker/src/api/android_saf_options.dart';
+import 'package:file_picker/src/api/windows_options.dart';
+import 'package:file_picker/src/api/linux_options.dart';
 import 'package:file_picker/src/platform/file_picker_platform_interface.dart';
 import 'package:file_picker/src/utils/file_picker_utils.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
@@ -46,6 +48,8 @@ class FilePickerWeb extends FilePickerPlatform {
     bool lockParentWindow = false,
     String? initialDirectory,
     AndroidSAFOptions? androidSafOptions,
+    WindowsOptions windowsOptions = const WindowsOptions(),
+    LinuxOptions linuxOptions = const LinuxOptions(),
   }) async {
     throw UnimplementedError('getDirectoryPath() has not been implemented.');
   }
@@ -65,6 +69,8 @@ class FilePickerWeb extends FilePickerPlatform {
     bool cancelUploadOnWindowBlur = true,
     int compressionQuality = 0,
     AndroidSAFOptions? androidSafOptions,
+    WindowsOptions windowsOptions = const WindowsOptions(),
+    LinuxOptions linuxOptions = const LinuxOptions(),
   }) async {
     FilePickerUtils.validateAllowedExtensions(type, allowedExtensions);
 
@@ -221,6 +227,8 @@ class FilePickerWeb extends FilePickerPlatform {
     required Uint8List bytes,
     Function(FilePickerStatus)? onFileLoading,
     bool lockParentWindow = false,
+    WindowsOptions windowsOptions = const WindowsOptions(),
+    LinuxOptions linuxOptions = const LinuxOptions(),
   }) async {
     if (bytes.isEmpty) {
       throw ArgumentError(