+++ lib/src/platform/windows/file_picker_windows.dart
@@ -9,6 +9,8 @@ import 'package:ffi/ffi.dart';
 import 'package:file_picker/src/api/file_picker_types.dart';
 import 'package:file_picker/src/api/file_picker_result.dart';
 import 'package:file_picker/src/api/android_saf_options.dart';
+import 'package:file_picker/src/api/windows_options.dart';
+import 'package:file_picker/src/api/linux_options.dart';
 
 import 'package:file_picker/src/api/exceptions.dart';
 import 'package:file_picker/src/platform/file_picker_platform_interface.dart';
@@ -38,7 +40,8 @@ class FilePickerWindows extends FilePickerPlatform {
     int compressionQuality = 0,
     bool cancelUploadOnWindowBlur = true,
     AndroidSAFOptions? androidSafOptions,
-    int? parentWindowHandle,
+    WindowsOptions windowsOptions = const WindowsOptions(),
+    LinuxOptions linuxOptions = const LinuxOptions(),
   }) async {
     final port = ReceivePort();
     await Isolate.spawn(
@@ -50,8 +53,8 @@ class FilePickerWindows extends FilePickerPlatform {
         type: type,
         allowedExtensions: allowedExtensions,
         allowMultiple: allowMultiple,
-        lockParentWindow: lockParentWindow,
-        parentWindowHandle: parentWindowHandle,
+        lockParentWindow: windowsOptions.lockParentWindow || lockParentWindow,
+        parentWindowHandle: windowsOptions.parentWindowHandle,
       ),
     );
     final fileNames = (await port.first) as List<String>?;
@@ -106,13 +109,14 @@ class FilePickerWindows extends FilePickerPlatform {
     bool lockParentWindow = false,
     String? initialDirectory,
     AndroidSAFOptions? androidSafOptions,
-    int? parentWindowHandle,
+    WindowsOptions windowsOptions = const WindowsOptions(),
+    LinuxOptions linuxOptions = const LinuxOptions(),
   }) async {
     return compute(_getDirectoryPathIsolate, {
       'dialogTitle': dialogTitle,
       'initialDirectory': initialDirectory,
-      'lockParentWindow': lockParentWindow,
-      'parentWindowHandle': parentWindowHandle,
+      'lockParentWindow': windowsOptions.lockParentWindow || lockParentWindow,
+      'parentWindowHandle': windowsOptions.parentWindowHandle,
     });
   }
 
@@ -156,13 +160,13 @@ class FilePickerWindows extends FilePickerPlatform {
         }
 
         try {
-          fileDialog.show(
-            lockParentWindow
-                ? (parentWindowHandle != null
-                    ? parentWindowHandle
-                    : GetForegroundWindow())
-                : null,
-          );
+          HWND? parentHwnd;
+          if (lockParentWindow) {
+            parentHwnd = parentWindowHandle != null
+                ? Pointer.fromAddress(parentWindowHandle) as HWND
+                : GetForegroundWindow();
+          }
+          fileDialog.show(parentHwnd);
         } on WindowsException catch (e) {
           if (e.hr == HRESULT_FROM_WIN32(ERROR_CANCELLED)) {
             return null;
@@ -198,7 +202,8 @@ class FilePickerWindows extends FilePickerPlatform {
     required Uint8List bytes,
     Function(FilePickerStatus)? onFileLoading,
     bool lockParentWindow = false,
-    int? parentWindowHandle,
+    WindowsOptions windowsOptions = const WindowsOptions(),
+    LinuxOptions linuxOptions = const LinuxOptions(),
   }) async {
     final port = ReceivePort();
     await Isolate.spawn(
@@ -210,9 +215,9 @@ class FilePickerWindows extends FilePickerPlatform {
         initialDirectory: initialDirectory,
         type: type,
         allowedExtensions: allowedExtensions,
-        lockParentWindow: lockParentWindow,
+        lockParentWindow: windowsOptions.lockParentWindow || lockParentWindow,
         confirmOverwrite: true,
-        parentWindowHandle: parentWindowHandle,
+        parentWindowHandle: windowsOptions.parentWindowHandle,
       ),
     );
     final savedFilePath = (await port.first) as String?;